добавили комментарии

+++ src/main/java/ru/zinyakova/kassa/entity/FullReceipt.java
@@ -2,6 +2,7 @@ package ru.zinyakova.kassa.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.ArrayList;
 
@@ -9,6 +10,6 @@ import java.util.ArrayList;
 public class FullReceipt {
     private Long id;
     private Date date;
-    private Long summa;
+    private BigDecimal summa;
     private ArrayList<FullReceiptItem> items = new ArrayList<>();
 }
