+++ src/main/java/ru/zinyakova/kassa/entity/Receipt.java
@@ -2,11 +2,12 @@ package ru.zinyakova.kassa.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Data
 public class Receipt {
     private Long id;
     private Date date;
-    private Long summa;
+    private BigDecimal summa; // сумма скидки
 }
