+++ src/main/java/ru/zinyakova/kassa/entity/ReceiptItem.java
@@ -2,11 +2,13 @@ package ru.zinyakova.kassa.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ReceiptItem {
     private Long id;
     private Long receiptId;
     private Long seatStatusId;
     private Long quantity;
-    private Long summa;
+    private BigDecimal summa; // сумма со скидкой
 }
